Update and rebase on WASI main.
As part of the move to modularize WASI, I've now reset the `main` branch
in this repository to be the same as that if the [`WASI` repository],
and this is now a PR on top of that.

This implements the directory structure described in WebAssembly/WASI#360,
with a standard directory, and a wasi-clocks directory under that.

[`WASI` repository]: https://github.com/WebAssembly/WASI/