Port the wasi-poll API to the new wit format.
This converts the wasi-poll repository to the new proposal template
format.

This redesigns the events and subscriptions types, to work with
interface types, and it separates monotonic from wall clocks, to
align with the current wasi-clocks proposal.

Wit is currently unable to express importing a type from another wit
file, so this proposal currently includes its own types for clocks
and file descriptors. In the future, these should be replaced with
types imported from the other wasi proposals.

This fixes #2 by eliminating the errno code returned from poll-oneoff.
It isn't needed, because poll-oneoff itself doesn't do any I/O so it
doesn't need to be able to fail.