Update to the latest wasi-proposal-repo format.
This updates to the latest wasi-proposal-repo format, which includes
updating to the latest wit-bindgen. That in turm includes temporarily
removing resources, though I've left in comments making it clear where
the resources will go.

This also converts `poll` from the `subscription` record API to a new
pollable API which allows it to avoid hard-coding the set of pollable
things in the API.